Fix open_basedir restriction issue
Enabling the Aviary plugin crashes Grav v1.5.0 / Admin v1.8.8.
This fix is a solution. I checked other non-crashing plugins 
and the current convention if about loading assets is using `plugin://` 
in the beginning of the path. Without this fix, the path was invalid 
and one `/` was missing. Then critical fragment was like 
`.../public_htmluser/plugins...` where the `/` should be before the `user`.
With proposed fix, the path it is being resolved correctly by Grav.